Admin Statistics: 1. Implemented all features and statistics. 2. Added styles to other statistics 3. Removed ngx-charts graphs (not needed)
Todo admin statistics:
1. Styling abandoned carts
2. Implement survey page and update the email to send the survey dynamic per user link

Other things:
1. Increased timeout in navbar, sometimes 1200 is not enough and the bug come back (created two shopping carts)
2.  Created email service that makes calls with ajax to external asmx.
3.  fixed not recived emails of users from firebase.

Abandoned Carts Logic:
If a registered user have a shopping-cart (and to let you know,there is only one shopping cart for registered user...), and the user had items in his cart, and email not sent to him for more then 2 weeks, and last thing is that the user didn't maked any action in the website for 2 days. then > show him as potential lost user.